requirements.txt
web3==6.0.0
requests==2.28.1
python-dotenv==1.0.0
joblib==1.2.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
pytest==7.2.2
flask==2.2.3
etherscan-python==2.1.0
matplotlib==3.6.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1